Bug Patch 1.3 ContentView.swift
if you tried removing a card in the array but did not set the stepper to 1 the app would crash, because what i added in bug patch 1.2 would subtract 1 so the removeCard variable would be set to negative one, but arrays starting at zero not negative one. so now if the remove card gets set as -1 one is added to make it zero again.